refactor(vm): split vm_string_regex_ops.rs into cohesive modules (§7-8)
Split the 3089-line vm_string_regex_ops.rs (over the 500-line limit)
into seven cohesive modules, all under 500 lines:

- vm_string_regex_ops.rs (414) - subst case/capture free fns + QuantKind
- vm_subst_exec.rs (483)       - exec_subst_op / non-destructive subst
- vm_subst_apply.rs (465)      - substitution range/spec/apply + interpolate
- vm_hyper_ops.rs (416)        - hyper op + quanthash + transliterate
- vm_hyper_func.rs (438)       - hyper-func op dispatch
- vm_meta_ops.rs (450)         - meta op + ZipIter enum/impl
- vm_flipflop_ops.rs (439)     - flip-flop + infix-func/fallback

All methods are impl Interpreter, so the split is visibility-only. The
module-level subst free fns (apply_subst_case_transforms,
expand_capture_refs) stay in the main module to preserve the
crate::vm::vm_string_regex_ops::* paths used by external callers; a few
cross-module private helpers and the QuantKind enum are bumped to
pub(super), and sibling modules glob in super::vm_string_regex_ops::*
where needed. The fn multiset is byte-identical. No behavior change.

Note: this file has regex/format string literals containing unbalanced
{ }, so methods were sliced by declaration-line ranges rather than brace
matching.
