+++ src/vm.rs
@@ -143,9 +143,13 @@ mod vm_control_ops;
 mod vm_data_ops;
 mod vm_dispatch_helpers;
 mod vm_env_helpers;
+pub(crate) mod vm_flipflop_ops;
 mod vm_helpers;
+pub(crate) mod vm_hyper_func;
 mod vm_hyper_method_ops;
+pub(crate) mod vm_hyper_ops;
 mod vm_hyper_race_parallel;
+pub(crate) mod vm_meta_ops;
 pub(crate) mod vm_method_dispatch;
 pub(crate) mod vm_misc_ops;
 mod vm_mixin_does_ops;
@@ -167,6 +171,8 @@ pub(crate) mod vm_smart_match;
 mod vm_smartmatch_ops;
 pub(crate) mod vm_stats;
 pub(crate) mod vm_string_regex_ops;
+pub(crate) mod vm_subst_apply;
+pub(crate) mod vm_subst_exec;
 mod vm_typedecl_ops;
 mod vm_value_helpers;
 mod vm_var_assign_ops;
